requirements.txt
astroid==1.6.3
attrs==17.4.0
cachetools==2.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
clarifai==2.0.33
configparser==3.5.0
dill==0.2.7.1
enum34==1.1.6
future==0.15.2
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
google-api-core==0.1.4
google-auth==1.4.1
google-cloud==0.32.0
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.1
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.1
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.0.1
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.10.1
httplib2==0.11.3
idna==2.6
isort==4.3.4
jsonschema==2.5.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.1.0
numpy==1.14.2
oauth2client==3.0.0
Pillow==2.9.0
pluggy==0.6.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.2.post1
psutil==5.4.3
py==1.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pylint==1.8.4
PySoundFile==0.9.0.post1
pytesseract==0.2.0
pytest==3.5.0
pytest-pylint==0.9.0
python-telegram-bot==10.0.1
pytz==2018.3
requests==2.13.0
rsa==3.4.2
scipy==1.0.1
six==1.11.0
urllib3==1.22
wrapt==1.10.11
